+++ Fluid/Values/DictionaryValue.cs
@@ -106,7 +106,7 @@ namespace Fluid.Values
             return true;
         }
 
-        public override double ToNumberValue()
+        public override decimal ToNumberValue()
         {
             return 0;
         }
